+++ sdata/migrations/013-kde-material-you-colors-stop.sh
@@ -39,7 +39,7 @@ migration_apply() {
   local ii_dir
   ii_dir="$(get_runtime_shell_dir)"
   ii_dir="${ii_dir:-${XDG_CONFIG_HOME:-$HOME/.config}/quickshell/ilmango}"
-  local wrapper_src="${ii_dir}/dots/.config/matugen/templates/kde/kde-material-you-colors-wrapper.sh"
+  local wrapper_src="${ii_dir}/defaults/matugen/templates/kde/kde-material-you-colors-wrapper.sh"
   local wrapper_dst="${XDG_CONFIG_HOME:-$HOME/.config}/matugen/templates/kde/kde-material-you-colors-wrapper.sh"
 
   # Copy the fixed wrapper from the repo